根据deltatime更改速度相关及得分组件更新

+++ src/Components/RocksCreatAndDestoryComponent.java
@@ -10,14 +10,14 @@ public class RocksCreatAndDestoryComponent extends ComponentSystem {
 
     private float rockCreateTime;
     private String rockName;
-    private float speedX;
-    private float speedY;
+    private int speedX;
+    private int speedY;
 
     private float timeLoop = 0;
     private int randomPositionY;
     private Random random = new Random();
 
-    public RocksCreatAndDestoryComponent(float rockCreateTime, String rockName, float speedX, float speedY) {
+    public RocksCreatAndDestoryComponent(float rockCreateTime, String rockName, int speedX, int speedY) {
         this.rockCreateTime = rockCreateTime;
         this.rockName = rockName;
         this.speedX = speedX;
